Added <C-y> and <C-p> to yank and paste to the system buffer and fugitive.vim
CTRL-C and CTRL-v buffer, not the middle click buffer in general that is the one
I prefer since the middle-click one is too easy to overwrite between the time
you select something and the moment you paste it.  Even though control-c and
control-v is available in less places than select middle click.