Reorganized CMC commands in bashrc
With the creation of an Ubuntu-14 directory for git, I moved the ssm commands so
that they are run on any CMC system and removed the aliases with the full paths
to the Ubuntu 12 version.

Also added a thing to start tmux on my machine if I'm not SSH'd.  It is
commented out for now because I'm not sure if i like it.