Changed behavior of arrow keys
When I'm browsing code, I like to use the mouse wheel to go up and down quickly
without jumping.  So that is why I have remapped the up-down arrow keys in
normal mode to keep the cursor in the same position in the screen whild
scrolling the text.

Also changed the messages that are sent.  All from the 1994 classic movie The
Shadow.