Added cdf function to go to where a file is
Use cdf to go to a directory contained anywhere in $PWD using the find command.

It will cd to the first result that find gets.

If the result is a directory, it will CD to that and if is a file, it will CD to
the directory containing that file.

Untested on symbolic links.