Solved problem with tmux and sourcing of bashrc
Exporting the PHILRC variables caused them to still be set in
subporcesses.  So for example, when launching TMUX, I want bashrc to be
sourced for each new pane/tab.  However since the parent process
of TMUX put these in the environment, the line in bash_profile didn't
do the right thing.

By changing them to shell variables this gives the desired change.