Moved some functions into an initutils file
The idea is that I want my functions and aliases to be sourced at the latest
time possible so that their settings override things set by other parts of the
shell startup process.

However other parts of the process use some functions like at_cmc or at_poly to
do some initial stuff.  So I split off the things that need to be sourced early
from the things that need to be sourced late.