Set default tmux teminal to xterm-256color
With the original value of `tmux-256color`, we had some weird things
happening.  In /etc/bashrc, PROMPT_COMMAND is set to
printf "\033]0;...\007"
or
printf "\033k...\033\\"
which the terminal emulator is supposed to pick up.

If the terminal emulator doesn't understand the \033k for example, then
teh output goes to the terminal.

The weirdness that motivated this change is not happening anymore now.